requirements.txt
python-telegram-bot==20.8
pytz
httpx
nest_asyncio
google-auth
google-auth-oauthlib
google-api-python-client
Pillow